+++ lib/flutter_staggered_grid_view.dart
@@ -4,6 +4,7 @@ export 'src/layouts/quilted.dart';
 export 'src/layouts/staired.dart';
 export 'src/layouts/woven.dart';
 export 'src/rendering/sliver_masonry_grid.dart';
+export 'src/rendering/sliver_simple_grid_delegate.dart';
 export 'src/widgets/aligned_grid_view.dart';
 export 'src/widgets/masonry_grid_view.dart';
 export 'src/widgets/sliver_aligned_grid.dart';